requirements.txt
cryptography==36.0.2
Django==4.0.3
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-mptt==0.13.4
django-quill-editor==0.1.40
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
Pillow==9.0.1
sslcommerz-lib==1.0
ujson==5.1.0
whitenoise==6.0.0
django-storages==1.12.3
boto3==1.24.2
django-nested-admin==3.4.0